pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Umit Silwal Khatri", email = "silwalumit@gmail.com" }
]
dependencies = [
    "click>=8.2.1",
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "langchain>=0.3.26",
    "langchain_community>=0.3.27",
    "sentence_transformers>=5.0.0",
    "faiss-cpu>=1.11.0.post1",
    "uvicorn[standard]>=0.35.0",
    "chromadb>=1.0.15",
    "langchain_chroma>=0.2.4",
    "langchain_huggingface>=0.3.0",
    "langchain-ollama>=0.3.5",
    "langgraph>=0.5.3",
    "langchain-openai>=0.3.28",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipykernel>=6.29.5",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/research_assistant"]

[tool.rye.scripts]
cli = { cmd = "python -m src.cmd.cli"} 
dev = { cmd = "python -m uvicorn src.cmd.main:app --loop uvloop --reload"}
